Merge C# Verification

+++ CSharpProject/src/Model/Ship.cs
@@ -2,16 +2,6 @@
 // <remarks> Deployment information is supplied to allow ships to be drawn.
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Globalization;
-using System.IO;
-using System.Linq;
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Security;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.VisualBasic;
 
 public class Ship
 {
@@ -96,7 +86,7 @@ public class Ship
         _tiles = new List<Tile>();
 
         // gets the ship size from the enumarator
-        _sizeOfShip = _shipName;
+        _sizeOfShip = (int)_shipName;
     }
 
     // <summary> Add tile adds the ship tile
